agent/requirements.txt
python-socketio[client]==5.11.2
websocket-client>=1.7.0
requests>=2.31.0
PyYAML>=6.0.1
